make filter reusable and add it to transactions
