+++ src/components/Filter/filterList.js
@@ -1,0 +1,7 @@
+const filterList = {
+  0: 2023,
+  1: 2022,
+  2: 2021,
+};
+
+export default filterList;
