+++ src/components/Transactions/Transactions.jsx
@@ -1,30 +1,8 @@
 import React from 'react';
+import Filter from '../Filter/Filter';
 import TransactionDate from '../TransactionDate/TransactionDate';
 import Tag from '../Tag/Tag';
-
-const expenses = [
-  {
-    id: 'tf1',
-    type: 'Food',
-    name: 'Supermarket',
-    amount: 35,
-    date: new Date(2023, 4, 29),
-  },
-  {
-    id: 'tr1',
-    type: 'Rent',
-    name: 'Apartment',
-    amount: 1500,
-    date: new Date(2023, 4, 29),
-  },
-  {
-    id: 'tt1',
-    type: 'Transport',
-    name: 'Bus',
-    amount: 2,
-    date: new Date(2023, 4, 26),
-  },
-];
+import expenses from './expenses';
 
 function Transactions() {
   const transactions = expenses.map((transaction) => (
@@ -44,7 +22,12 @@ function Transactions() {
     </li>
   ));
 
-  return <ul className="divide-y divide-gray-100">{transactions}</ul>;
+  return (
+    <section>
+      <Filter title="Transactions" options="Year" />
+      <ul className="divide-y divide-gray-100">{transactions}</ul>
+    </section>
+  );
 }
 
 export default Transactions;
