+++ src/components/Transactions/expenses.js
@@ -1,0 +1,25 @@
+const expenses = [
+  {
+    id: 'tf1',
+    type: 'Food',
+    name: 'Supermarket',
+    amount: 35,
+    date: new Date(2023, 4, 29),
+  },
+  {
+    id: 'tr1',
+    type: 'Rent',
+    name: 'Apartment',
+    amount: 1500,
+    date: new Date(2023, 4, 29),
+  },
+  {
+    id: 'tt1',
+    type: 'Transport',
+    name: 'Bus',
+    amount: 2,
+    date: new Date(2023, 4, 26),
+  },
+];
+
+export default expenses;
